working on delete!

+++ Final/models/celebrity_model.class.php
@@ -261,7 +261,15 @@ class CelebrityModel
         echo "<div align='center'><h1 style='color:green'><b>You have successfully added a new celebrity! :)</b></h1></div>";
     }
 
-    public function delete_celeb(){
+    public function delete_celeb($celeb_id){
+
+        if (!filter_has_var(INPUT_GET, 'celeb_id')) {
+            echo "Deletion cannot continue since there were problems retrieving product id";
+            die();
+        }
+        //this makes the variable id from the product id value and sanitizes/filters it
+        $celeb_id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
+
         $sql = "DELETE FROM celebrity WHERE celebrity . celeb_id = $celeb_id";
 
         //executes the query
@@ -272,6 +280,9 @@ class CelebrityModel
 
         // close the connection.
         $this->dbConnection->close();
+
+        //display a confirmation message
+        echo "<p>You have successfully deleted the item.</p><br>";
     }
